[core] Move {clip,merge}_lines to layout directory
