+++ src/mbgl/layout/clip_lines.hpp
