+++ src/mbgl/layout/merge_lines.hpp
